Revert renaming of PRBs settings
