+++ client/entrypoints/payment-request-settings/__tests__/payment-request-button-preview.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { useStripe } from '@stripe/react-stripe-js';
-import ExpressCheckoutButtonPreview from '../express-checkout-button-preview';
+import PaymentRequestsButtonPreview from '../payment-request-button-preview';
 import { shouldUseGooglePayBrand } from '../utils/utils';
 
 // We need to mock the actual module being used by `<Notice />` in the `@wordpress/components` module
@@ -30,7 +30,7 @@ jest.mock( 'wcstripe/data', () => ( {
 	usePaymentRequestButtonTheme: jest.fn().mockReturnValue( [ 'dark' ] ),
 } ) );
 
-describe( 'ExpressCheckoutButtonPreview', () => {
+describe( 'PaymentRequestsButtonPreview', () => {
 	const canMakePaymentMock = jest.fn();
 
 	beforeEach( () => {
@@ -50,7 +50,7 @@ describe( 'ExpressCheckoutButtonPreview', () => {
 	it( 'displays Google Chrome and Google Pay when page is in Safari', async () => {
 		shouldUseGooglePayBrand.mockReturnValue( false );
 
-		render( <ExpressCheckoutButtonPreview /> );
+		render( <PaymentRequestsButtonPreview /> );
 
 		expect(
 			await screen.findByText(
@@ -63,7 +63,7 @@ describe( 'ExpressCheckoutButtonPreview', () => {
 	it( 'displays Safari Apple Pay when page is in Google Chrome', async () => {
 		shouldUseGooglePayBrand.mockReturnValue( true );
 
-		render( <ExpressCheckoutButtonPreview /> );
+		render( <PaymentRequestsButtonPreview /> );
 
 		expect(
 			await screen.findByText(
@@ -76,7 +76,7 @@ describe( 'ExpressCheckoutButtonPreview', () => {
 	it( 'displays an info notice if stripe is falsy', async () => {
 		useStripe.mockReturnValue( null );
 
-		render( <ExpressCheckoutButtonPreview /> );
+		render( <PaymentRequestsButtonPreview /> );
 
 		expect(
 			screen.queryByText( 'Stripe button mock' )
@@ -90,7 +90,7 @@ describe( 'ExpressCheckoutButtonPreview', () => {
 
 	it( 'displays an info notice if stripe fails to load', async () => {
 		canMakePaymentMock.mockResolvedValue( null );
-		render( <ExpressCheckoutButtonPreview /> );
+		render( <PaymentRequestsButtonPreview /> );
 
 		expect(
 			await screen.findByText(
@@ -103,7 +103,7 @@ describe( 'ExpressCheckoutButtonPreview', () => {
 	} );
 
 	it( 'displays the payment button when stripe is loaded', async () => {
-		render( <ExpressCheckoutButtonPreview /> );
+		render( <PaymentRequestsButtonPreview /> );
 
 		expect(
 			await screen.findByText( 'Stripe button mock' )
