+++ client/entrypoints/payment-request-settings/__tests__/payment-request-settings-locations.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import ExpressCheckoutSettingsSection from '../express-checkout-settings-section';
-import ExpressCheckoutButtonPreview from '../express-checkout-button-preview';
+import PaymentRequestsSettingsSection from '../payment-request-settings-section';
+import PaymentRequestButtonPreview from '../payment-request-button-preview';
 import {
 	usePaymentRequestEnabledSettings,
 	usePaymentRequestLocations,
@@ -24,8 +24,8 @@ jest.mock( 'wcstripe/data/account-keys/hooks', () => ( {
 	useAccountKeysTestPublishableKey: jest.fn().mockReturnValue( [ '' ] ),
 } ) );
 
-jest.mock( '../express-checkout-button-preview' );
-ExpressCheckoutButtonPreview.mockImplementation( () => '<></>' );
+jest.mock( '../payment-request-button-preview' );
+PaymentRequestButtonPreview.mockImplementation( () => '<></>' );
 
 jest.mock( '../utils/utils', () => ( {
 	getPaymentRequestData: jest.fn().mockReturnValue( {
@@ -81,7 +81,7 @@ describe( 'PaymentRequestsSettingsSection', () => {
 	} );
 
 	it( 'should enable express checkout locations when express checkout is enabled', () => {
-		render( <ExpressCheckoutSettingsSection /> );
+		render( <PaymentRequestsSettingsSection /> );
 
 		const [ checkoutCheckbox, productPageCheckbox, cartCheckbox ] =
 			screen.getAllByRole( 'checkbox' );
@@ -106,7 +106,7 @@ describe( 'PaymentRequestsSettingsSection', () => {
 			)
 		);
 
-		render( <ExpressCheckoutSettingsSection /> );
+		render( <PaymentRequestsSettingsSection /> );
 
 		// Uncheck each checkbox, and verify them what kind of action should have been called
 		await userEvent.click( screen.getByText( 'Product page' ) );
@@ -140,7 +140,7 @@ describe( 'PaymentRequestsSettingsSection', () => {
 			)
 		);
 
-		render( <ExpressCheckoutSettingsSection /> );
+		render( <PaymentRequestsSettingsSection /> );
 
 		await userEvent.click( screen.getByText( 'Cart' ) );
