+++ client/entrypoints/payment-request-settings/__tests__/payment-request-settings.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import ExpressCheckoutSettingsSection from '../express-checkout-settings-section';
-import ExpressCheckoutButtonPreview from '../express-checkout-button-preview';
+import PaymentRequestsSettingsSection from '../payment-request-settings-section';
+import PaymentRequestButtonPreview from '../payment-request-button-preview';
 import {
 	usePaymentRequestEnabledSettings,
 	usePaymentRequestLocations,
@@ -27,8 +27,8 @@ jest.mock( 'wcstripe/data/account-keys/hooks', () => ( {
 } ) );
 jest.mock( '@woocommerce/blocks-checkout', () => {}, { virtual: true } );
 
-jest.mock( '../express-checkout-button-preview' );
-ExpressCheckoutButtonPreview.mockImplementation( () => '<></>' );
+jest.mock( '../payment-request-button-preview' );
+PaymentRequestButtonPreview.mockImplementation( () => '<></>' );
 
 jest.mock( '../utils/utils', () => ( {
 	getPaymentRequestData: jest.fn().mockReturnValue( {
@@ -82,7 +82,7 @@ describe( 'PaymentRequestsSettingsSection', () => {
 	} );
 
 	it( 'renders settings with defaults', () => {
-		render( <ExpressCheckoutSettingsSection /> );
+		render( <PaymentRequestsSettingsSection /> );
 
 		// confirm settings headings.
 		expect(
@@ -125,7 +125,7 @@ describe( 'PaymentRequestsSettingsSection', () => {
 		] );
 		usePaymentRequestEnabledSettings.mockReturnValue( [ true, jest.fn() ] );
 
-		render( <ExpressCheckoutSettingsSection /> );
+		render( <PaymentRequestsSettingsSection /> );
 
 		expect( setButtonTypeMock ).not.toHaveBeenCalled();
 		expect( setButtonSizeMock ).not.toHaveBeenCalled();
