+++ webpack.config.js
@@ -113,13 +113,13 @@ module.exports = {
 	},
 	entry: {
 		index: './client/blocks/index.js',
+		'payment-requests-settings':
+			'./client/entrypoints/payment-request-settings/index.js',
 		'upe-classic': './client/classic/upe/index.js',
 		'upe-blocks': './client/blocks/upe/index.js',
 		'upe-settings': './client/settings/index.js',
 		'payment-gateways': './client/entrypoints/payment-gateways/index.js',
 		'express-checkout': './client/entrypoints/express-checkout/index.js',
-		'express-checkout-settings':
-			'./client/entrypoints/express-checkout-settings/index.js',
 		'amazon-pay-settings':
 			'./client/entrypoints/amazon-pay-settings/index.js',
 	},
